requirements.txt
discord.py[voice]
pillow
yt_dlp